+++ src/main/java/waktfolio/domain/repository/like/MemberLikeCustomRepository.java
@@ -1,8 +1,9 @@
 package waktfolio.domain.repository.like;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface MemberLikeCustomRepository {
-    Long countAddCountByContentId(UUID contentId);
+    BigDecimal countAddCountByContentId(UUID contentId);
     Long countAddCountByMemberId(UUID memberId);
 }
