+++ src/main/java/waktfolio/domain/repository/view/ContentViewCustomRepository.java
@@ -1,8 +1,10 @@
 package waktfolio.domain.repository.view;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.UUID;
 
 public interface ContentViewCustomRepository {
-    Long sumAddSumByContentId(UUID contentId);
+    BigDecimal sumAddSumByContentId(UUID contentId);
     Long sumAddSumByMemberId(UUID memberId);
 }
